requirements.txt
flask==3.0.0
anthropic==0.40.0
python-dotenv==1.0.0
Pillow==10.2.0
